ext/docs: refactoring code
