+++ src/models/model_setup.py
@@ -1,7 +1,5 @@
 """Set up the model, copy the files, get the names."""
-from typing import List
 import os
-import pandas as pd
 from omegaconf import DictConfig
 from src.constants import (
     OCEAN_RUN_PATH,
@@ -12,52 +10,7 @@ from src.constants import (
     MODEL_NAMES,
     VAR_DICT,
 )
-
-
-def mems_to_df(mem_list: List[str]) -> pd.DataFrame:
-    """
-    Turn a list of mems into a dataframe of inputs.
-
-    Args:
-        mem_list (List[str]): List of mem to turn into corresponding
-            dataframe of inputs.
-
-    Returns:
-        pd.DataFrame: A dataframe
-
-    Example:
-        Work out what inputs a list of runs got::
-
-            mems_to_df(["EEEE", "CCCC", "66E6"])
-    """
-    results_lol = []
-    for i in VAR_DICT:
-        results_lol.append([])
-    for i, mem in enumerate(mem_list):
-        for j in VAR_DICT:
-            if len(mem) <= j:
-                results_lol[j].append(MODEL_NAMES["E"])
-            else:
-                results_lol[j].append(MODEL_NAMES[mem[j]])
-    return pd.DataFrame(
-        data={VAR_DICT[i]: results_lol[i] for i in range(len(results_lol))},
-        index=mem_list,
-    )
-
-
-def mem_to_dict(mem: str) -> dict:
-    """
-    Single mem variable to dictionary of inputs.
-
-    Uses logic in `mems_to_df`.
-
-    Args:
-        mem (str): the mem input e.g "EEEE"
-
-    Returns:
-        dict: dictionary of inputs, e.g.
-    """
-    return {var: input_d[mem] for (var, input_d) in mems_to_df([mem]).to_dict().items()}
+from src.mem_to_input import mem_to_dict
 
 
 class ModelSetup:
